refactor(seo): extract lens-SEO helpers, share JsonLd component
Two changes in one commit because they overlap on the same files:

1. Pulled `buildLensDescription` and `buildLensProductSchema` out of the
   lens detail page.tsx into `src/lib/lens-seo.ts`. The page now reads
   as thin orchestration — resolve params, look up lens, call two
   helpers, return Metadata / render JSX. Inlined `isPrime` and
   `mountLabel` are replaced with `isZoom(lens)` (already in
   `src/lib/lens.ts`) and a new `mountSeoLabel(mount)` helper next to
   the existing `mountToUrlSegment` in `src/lib/mount.ts`. Net: ~140
   lines of business logic out of the page file.

2. New `<JsonLd>` component encapsulates the script-tag pattern and
   adds a defensive `<` → `<` escape. Without that escape, any
   `</script>` substring inside a JSON string value would prematurely
   close the script tag at the byte-stream level (HTML parsers tokenize
   script-data without knowing about JSON quoting). Our current data
   doesn't contain `<` anywhere, but the escape costs ~5 chars and
   removes the future-bug surface area. Both SiteJsonLd (site-level
   Organization + WebSite graph) and the lens detail page (Product
   schema) route through it, so the escape stays consistent.
